+++ Threeyes/SDK/Scripts/Mod/BuiltIn/ToDelete/Animation/AC_SOCharacterAnimatorControllerConfig.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using Threeyes.Config;
-using Threeyes.Steamworks;
-using UnityEngine;
-
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_Root_BuiltIn + "Animation/CharacterAnimatorController", fileName = "CharacterAnimatorControllerConfig")]
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_SOCharacterAnimatorControllerConfig : SOCharacterAnimatorControllerConfig { }