+++ Threeyes/SDK/Scripts/Mod/BuiltIn/ToDelete/Transform/AC_LookAtCamera.cs
@@ -1,6 +1,0 @@
-using Threeyes.Config;
-using Threeyes.Steamworks;
-using UnityEngine;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_LookAtCamera : LookAtCamera { }