+++ Threeyes/SDK/Scripts/Mod/BuiltIn/ToDelete/Transform/AC_SOObjectMovement_FollowTargetConfig.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using Threeyes.Config;
-using Threeyes.Steamworks;
-using UnityEngine;
-
-[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_Root_BuiltIn + "Transform/ObjectMovement/FollowTarget", fileName = "ObjectMovement_FollowTargetConfig")]
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_SOObjectMovement_FollowTargetConfig : SOObjectMovement_FollowTargetConfig { }