+++ Threeyes/SDK/Scripts/Mod/BuiltIn/ToDelete/Video/AC_VideoPlayerHelper.cs
@@ -1,7 +1,0 @@
-using UnityEngine.Video;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_VideoPlayerHelper : ComponentHelperBase<VideoPlayer>
-{
-
-}