+++ Threeyes/SDK/Scripts/Mod/BuiltIn/ToDelete/VisualEffect/AC_VisualEffectHelperGroup.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.VFX;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_VisualEffectHelperGroup : VisualEffectHelperGroup{}