+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/Clock/AC_ClockController.cs
@@ -1,8 +1,0 @@
-using System;
-using UnityEngine;
-using DG.Tweening;
-using Threeyes.Config;
-using Threeyes.Steamworks;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_ClockController : ClockController { }