+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/Creeper/AC_CreeperAudioVisualizer.cs
@@ -1,9 +1,0 @@
-using Threeyes.Steamworks;
-/// <summary>
-/// Creeper response To Audio
-/// 
-/// 功能：
-/// -通过序号确定指定响应音频的Leg
-/// </summary>
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_CreeperAudioVisualizer : CreeperAudioVisualizer { }