+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/Creeper/AC_CreeperLegController.cs
@@ -1,3 +1,0 @@
-using Threeyes.Steamworks;
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_CreeperLegController : CreeperLegController { }