+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/Creeper/AC_CreeperTransformController.cs
@@ -1,5 +1,0 @@
-
-using Threeyes.Steamworks;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_CreeperTransformController : CreeperTransformController { }