+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/ObjectGenerator/AC_GenerableObject.cs
@@ -1,5 +1,0 @@
-using Threeyes.Steamworks;
-using UnityEngine;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public abstract class AC_GenerableObject<TConfigInfo> : GenerableObject<TConfigInfo> { }