+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/ZibraLiquids/AC_ZibraLiquidAdvancedRenderParametersController.cs
@@ -1,9 +1,0 @@
-using Threeyes.Steamworks;
-/// <summary>
-///
-///
-/// Warning:
-/// 1.This related Component is rather complex, normally you don't need to expose the config to user.
-/// </summary>
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_ZibraLiquidAdvancedRenderParametersController : ZibraLiquidAdvancedRenderParametersController { }