+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/ZibraLiquids/AC_ZibraLiquidEmitterController.cs
@@ -1,4 +1,0 @@
-using Threeyes.Steamworks;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public class AC_ZibraLiquidEmitterController : ZibraLiquidEmitterController { }