+++ Threeyes/SDK/Scripts/Mod/Feature/ToDelete/ZibraLiquids/IAC_ZibraLiquidController_SettingHandler.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using Threeyes.Steamworks;
-using UnityEngine;
-
-[System.Obsolete("Use the class without AC_ prefix instead!", true)]
-public interface IAC_ZibraLiquidController_SettingHandler: IZibraLiquidController_SettingHandler
-{
-}